pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "process"
version = "0.1.0"
description = "Ainu corpora processing tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ainconv>=0.8.0",
    "datasets>=2.16.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.9.2",
    "nltk>=3.9.1",
    "openai>=1.57.3",
    "parsel>=1.9.1",
    "pdf2image>=1.17.0",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.0",
    "scipy>=1.14.1",
    "seaborn>=0.13.2",
    "starcc>=0.0.5",
    "utils",
    "wiktionary-dump-extractor",
    "wiktionary",
    "wordcloud>=1.9.4",
    "networkx>=3.4.2",
]

[project.scripts]
compare-wiktionary-lemmas = "dictionary.compare_wiktionary_lemmas:main"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["dictionary"]

[tool.uv.workspace]
members = ["utils"]

[tool.uv.sources]
utils = { workspace = true }
wiktionary-dump-extractor = { path = "wiktionary-dump-extractor" }
wiktionary = { git = "https://github.com/mkpoli/wiktionary-ja.git", subdirectory = "packages/wiktionary", rev = "b7e4dd82a083baefd5614d7569745c2e71834981" }
